+++ src/extension/module.js
@@ -420,6 +420,7 @@
       return true;
     } else {
       return [
+        'json',
         'jpg',
         'jpeg',
         'png',
@@ -432,30 +433,21 @@
   /**
    * Recognizes a SharePoint URL.
    * @private
-   * @param {Sidekick} sk The sidekick
    * @param {URL} url The URL
    * @returns {boolean} {@code true} if URL is SharePoint, else {@code false}
    */
-  function isSharePoint(sk, url) {
-    const { config } = sk;
-    const { mountpoint } = config;
-    const isSPMountpoint = typeof mountpoint === 'object'
-      ? mountpoint.contentSourceType === 'onedrive'
-      : /\w+\.sharepoint.com$/.test(mountpoint);
-
-    return /\w+\.sharepoint.com$/.test(url.host)
-      || (isSPMountpoint && url.host === new URL(mountpoint).host); // check custom sharepoint host
+  function isSharePoint(url) {
+    return /\w+\.sharepoint.com$/.test(url.host);
   }
 
   /**
    * Recognizes a SharePoint document management URL.
    * @private
-   * @param {Sidekick} sk The sidekick
    * @param {URL} url The URL
    * @returns {boolean} {@code true} if URL is SharePoint DM, else {@code false}
    */
-  function isSharePointDM(sk, url) {
-    return isSharePoint(sk, url)
+  function isSharePointDM(url) {
+    return isSharePoint(url)
       && (url.pathname.endsWith('/Forms/AllItems.aspx')
       || url.pathname.endsWith('/onedrive.aspx'));
   }
@@ -463,12 +455,11 @@
   /**
    * Recognizes a SharePoint folder URL.
    * @private
-   * @param {Sidekick} sk The sidekick
    * @param {URL} url The URL
    * @returns {boolean} {@code true} if URL is SharePoint folder, else {@code false}
    */
-  function isSharePointFolder(sk, url) {
-    if (isSharePointDM(sk, url)) {
+  function isSharePointFolder(url) {
+    if (isSharePointDM(url)) {
       const docPath = new URLSearchParams(url.search).get('id');
       const dotIndex = docPath?.split('/').pop().indexOf('.');
       return [-1, 0].includes(dotIndex); // dot only allowed as first char
@@ -482,9 +473,9 @@
    * @param {URL} url The URL
    * @returns {boolean} {@code true} if URL is SharePoint editor, else {@code false}
    */
-  function isSharePointEditor(sk, url) {
+  function isSharePointEditor(url) {
     const { pathname, search } = url;
-    return isSharePoint(sk, url)
+    return isSharePoint(url)
       && pathname.match(/\/_layouts\/15\/[\w]+.aspx$/)
       && search.includes('sourcedoc=');
   }
@@ -495,8 +486,8 @@
    * @param {URL} url The URL
    * @returns {boolean} {@code true} if URL is SharePoint viewer, else {@code false}
    */
-  function isSharePointViewer(sk, url) {
-    if (isSharePointDM(sk, url)) {
+  function isSharePointViewer(url) {
+    if (isSharePointDM(url)) {
       const docPath = new URLSearchParams(url.search).get('id');
       const dotIndex = docPath?.split('/').pop().lastIndexOf('.');
       return dotIndex > 0; // must contain a dot
@@ -515,6 +506,7 @@
       glob = '**';
     }
     const reString = glob
+      .replace('.', '\\.') // don't match every char, just real dots
       .replace(/\*\*/g, '_')
       .replace(/\*/g, '[0-9a-z-.]*')
       .replace(/_/g, '.*');
@@ -738,6 +730,21 @@
       return new URL(resource, origin);
     }
     return url;
+  }
+
+  /**
+   * Checks if the location has changed.
+   * @private
+   * @param {Sidekick} sk The sidekick
+   * @returns {boolean} {@code true} if location changed, else {@code false}
+   */
+  function isNewLocation(sk) {
+    const { location } = sk;
+    const $test = document.getElementById('sidekick_test_location');
+    if ($test) {
+      return $test.value !== location.href;
+    }
+    return window.location.href !== location.href;
   }
 
   /**
@@ -1538,7 +1545,7 @@
       const { path, type } = item;
       const nameParts = path.split('.');
       let [file, ext] = nameParts;
-      if (isSharePointFolder(sk, sk.location) && ext === 'docx') {
+      if (isSharePointFolder(sk.location) && ext === 'docx') {
         // omit docx extension on sharepoint
         ext = '';
       }
@@ -1561,7 +1568,7 @@
 
     const getBulkSelection = () => {
       const { location } = sk;
-      if (isSharePointFolder(sk, location)) {
+      if (isSharePointFolder(location)) {
         const isGrid = document.querySelector('div[class~="ms-TilesList"]');
         return [...document.querySelectorAll('#appRoot [role="presentation"] div[aria-selected="true"]')]
           .filter((row) => !row.querySelector('img')?.getAttribute('src').includes('/foldericons/')
@@ -1570,7 +1577,7 @@
           .map((row) => ({
             type: isGrid
               ? row.querySelector(':scope i[aria-label]')?.getAttribute('aria-label').trim()
-              : new URL(row.querySelector('img').getAttribute('src'), sk.location.href).pathname.split('/').slice(-1)[0].split('.')[0],
+              : new URL(row.querySelector('img')?.getAttribute('src'), sk.location.href).pathname.split('/').slice(-1)[0].split('.')[0],
             path: isGrid
               ? row.querySelector('div[data-automationid="name"]').textContent.trim()
               : row.querySelector('button')?.textContent.trim(),
@@ -1585,14 +1592,6 @@
               || row.querySelector(':scope > div > div > div:nth-of-type(4)').textContent.trim(), // grid layout
           }));
       }
-    };
-
-    const isChangedUrl = () => {
-      const $test = document.getElementById('sidekick_test_location');
-      if ($test) {
-        return $test.value !== sk.location.href;
-      }
-      return window.location.href !== sk.location.href;
     };
 
     const updateBulkInfo = () => {
@@ -1612,15 +1611,15 @@
       }
       // show/hide bulk buttons
       const filesSelected = sel.length > 0;
-      if (isChangedUrl()) {
-        // refresh location
-        sk.location = getLocation();
-      }
       ['preview', 'publish', 'copy-urls'].forEach((action) => {
         const pluginId = `bulk-${action}`;
         const plugin = sk.get(pluginId);
-        const customShowPlugin = (sk.customPlugins[action]?.condition || (() => true))(sk);
-        plugin.classList[filesSelected && customShowPlugin ? 'remove' : 'add']('hlx-sk-hidden');
+        let customShow = true;
+        const customPlugin = sk.customPlugins[action];
+        if (customPlugin) {
+          customShow = customPlugin.condition(sk);
+        }
+        plugin.classList[filesSelected && customShow ? 'remove' : 'add']('hlx-sk-hidden');
       });
       // update copy url button texts based on selection size
       ['', 'preview', 'live', 'prod'].forEach((env) => {
@@ -1760,7 +1759,7 @@
       callback: (sidekick) => {
         const { location } = sk;
         const listener = () => window.setTimeout(() => updateBulkInfo(sidekick), 100);
-        const rootEl = document.querySelector(isSharePointFolder(sk, location) ? '#appRoot' : 'body');
+        const rootEl = document.querySelector(isSharePointFolder(location) ? '#appRoot' : 'body');
         if (rootEl) {
           rootEl.addEventListener('click', listener);
           rootEl.addEventListener('keyup', listener);
@@ -1862,8 +1861,6 @@
         },
       });
     });
-
-    updateBulkInfo();
   }
 
   /**
@@ -2047,12 +2044,12 @@
             container: containerId,
           };
           sk.customPlugins[plugin.id] = plugin;
-          // check and remove existing plugin
-          const existingPlugin = sk.plugins[plugin.id];
-          if (existingPlugin) {
-            if (sk.get(plugin.id) && !condition(sk)) {
-              sk.remove(id);
-            }
+          // check default plugin
+          const defaultPlugin = sk.plugins[plugin.id];
+          if (defaultPlugin) {
+            // extend default condition
+            const { condition: defaultCondition } = defaultPlugin;
+            defaultPlugin.condition = (s) => defaultCondition(s) && condition(s);
           } else {
             // add custom plugin
             sk.add(plugin);
@@ -2512,18 +2509,15 @@
   }
 
   /**
-   * Registers a plugin for re-evaluation if it should be shown or hidden,
-   * and if its button should be enabled or disabled.
+   * Checks existing plugins based on the status of the current resource.
    * @private
    * @param {Sidekick} sk The sidekick
-   * @param {Plugin} plugin The plugin configuration
-   * @param {HTMLElement} $plugin The plugin
-   * @returns {HTMLElement} The plugin or {@code null}
    */
-  function registerPlugin(sk, plugin, $plugin) {
-    // re-evaluate plugin when status fetched
-    sk.addEventListener('statusfetched', () => {
-      const { status } = sk;
+  function checkPlugins(sk) {
+    const { status, plugins, pluginContainer } = sk;
+    Object.keys(plugins).forEach((id) => {
+      const plugin = plugins[id];
+      const $plugin = sk.get(id);
       if (typeof plugin.condition === 'function') {
         if ($plugin && !plugin.condition(sk)) {
           // plugin exists but condition now false
@@ -2551,24 +2545,12 @@
         }
       }
     });
-    if (typeof plugin.callback === 'function') {
-      plugin.callback(sk, $plugin);
-    }
-    return $plugin || null;
-  }
-
-  /**
-   * Checks existing plugins based on the status of the current resource.
-   * @private
-   * @param {Sidekick} sk The sidekick
-   */
-  function checkPlugins(sk) {
-    sk.pluginContainer.classList.remove('hlx-sk-concealed');
+    pluginContainer.classList.remove('hlx-sk-concealed');
     window.setTimeout(() => {
-      if (!sk.pluginContainer.querySelector(':scope div.plugin')) {
+      if (!pluginContainer.querySelector(':scope div.plugin')) {
         // add empty text
-        sk.pluginContainer.innerHTML = '';
-        sk.pluginContainer.append(createTag({
+        pluginContainer.innerHTML = '';
+        pluginContainer.append(createTag({
           tag: 'span',
           text: i18n(sk, 'plugins_empty'),
           attrs: {
@@ -2943,6 +2925,15 @@
 
       // collapse dropdowns when document is clicked
       document.addEventListener('click', () => collapseDropdowns(this));
+      // listen to URL changes
+      window.setInterval(() => {
+        if (isNewLocation(this)) {
+          this.fetchStatus(true);
+        }
+      }, 500);
+      window.addEventListener('popstate', () => {
+        this.fetchStatus(true);
+      });
     }
 
     /**
@@ -3134,7 +3125,7 @@
       plugin.isShown = typeof plugin.condition === 'undefined'
           || (typeof plugin.condition === 'function' && plugin.condition(this));
       if (!plugin.isShown) {
-        return registerPlugin(this, plugin, null);
+        return null;
       }
 
       // find existing plugin
@@ -3222,7 +3213,11 @@
       if (typeof plugin.advanced === 'function' && plugin.advanced(this)) {
         $plugin.classList.add('hlx-sk-advanced-only');
       }
-      return registerPlugin(this, plugin, $plugin);
+      // callback
+      if (typeof plugin.callback === 'function') {
+        plugin.callback(this, $plugin);
+      }
+      return $plugin;
     }
 
     /**
@@ -3252,12 +3247,15 @@
      * @returns {boolean} <code>true</code> if editor URL, else <code>false</code>
      */
     isEditor() {
-      const { location } = this;
+      const { config, location } = this;
       const { host } = location;
-      if (isSharePointEditor(this, location) || isSharePointViewer(this, location)) {
+      if (isSharePointEditor(location) || isSharePointViewer(location)) {
         return true;
       }
       if (host === 'docs.google.com') {
+        return true;
+      }
+      if (config.mountpoint && new URL(config.mountpoint).host === host && !this.isAdmin()) {
         return true;
       }
       return false;
@@ -3269,7 +3267,7 @@
      */
     isAdmin() {
       const { location } = this;
-      return isSharePointFolder(this, location) || location.host === 'drive.google.com';
+      return isSharePointFolder(location) || location.host === 'drive.google.com';
     }
 
     /**